+++ src/animemangadb/__init__.py
@@ -1,0 +1,1 @@
+"""AnimeMangaDB Platform v3 backend package."""